+++ gui/extension/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - BUG#38623232 Broken hyperlink for MSG certificate setup
 - BUG#38603422 Unexpected MRS SQL parsing error for field names that match certain keywords
+- BUG#35420780 mrs.configure() changes global session state
 -
 
 ## Changes in 1.19.19+9.5.0